requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.3.0
Flask==2.3.3
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pytz==2025.2
six==1.17.0
Werkzeug==3.1.3
python-dotenv==1.1.1
flask-marshmallow==1.3.0
bcrypt==5.0.0
PyJWT==2.10.1
redis==6.4.0
black==25.9.0
itsdangerous==2.2.0
PyMySQL==1.1.2
coverage==7.11.0
flasgger==0.9.7.1
